Propagate package renaming
