+++ hazelcast/src/main/java/com/hazelcast/memory/MemoryManagerBean.java
@@ -16,7 +16,7 @@
 
 package com.hazelcast.memory;
 
-import com.hazelcast.internal.memory.MemoryAccessor;
+import com.hazelcast.spi.memory.MemoryAccessor;
 
 /**
  * Simp]e implementation of {@link MemoryManager} which aggregates the
