+++ hazelcast/src/test/java/com/hazelcast/spi/impl/hashslot/HashSlotArray16byteKeyNoValueTest.java
@@ -16,7 +16,7 @@
 
 package com.hazelcast.spi.impl.hashslot;
 
-import com.hazelcast.memory.HeapMemoryManager;
+import com.hazelcast.spi.impl.memory.HeapMemoryManager;
 import com.hazelcast.memory.MemoryManager;
 import com.hazelcast.spi.hashslot.HashSlotArray16byteKey;
 import com.hazelcast.spi.hashslot.HashSlotCursor16byteKey;
